+++ CHANGES.rst
@@ -21,6 +21,7 @@
 - Removed a lot of unused code
 - Water Balance: updated Water Balance tool (#856, #868, #855)
 - Statistics: added preset "Water on street duration" (#845)
+- Statistics: Change style for water on street duration preset (#872)
 
 
 3.1.5 (2023-06-21)